add task sidecar metadata
